+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Protocol for unmanaged resources, defining a method to free the resource
+- Initialize GLFW
+- Create windows
+- Set callbacks on most types of input
+- Hide and show windows
 
 [Unreleased]: https://github.com/IGJoshua/s-expresso/compare/master..develop